chore(sweep): more dead code + first batch of test moves (ssz, javm-cap)
Dead code (continuing the previous sweep)
-----------------------------------------
`javm-recompiler-x86/src/asm.rs`: dropped six unused helpers and the
const that backed them — `emit_u64`, `rex_w`, `rex_opt`, `modrm_rr`,
`modrm_disp`, `modrm_sib_base_index`, and `CALLER_SAVED`. All had no
callers; the InstBuf-based variants are the live versions.

`nub-arch-x86/src/call_loop.rs`: removed the dead `ERR_DERIVE_PUBLISH`
constant. The publish-OOM path it covered no longer exists
(`publish_transient_instance` is infallible). Left a comment so
existing error codes don't shift.

`nub-arch-x86/src/paging.rs`: dropped the now-orphan `write_cr3()`
(only called by the deleted `flush_tlb()`).

Test reorganisation (round 1)
-----------------------------
Continuing the convention shift: black-box tests live in
`<crate>/tests/`, inline `mod tests` is reserved for cases that need
module-private access.

Moved out of `src/`:

- `ssz/src/{bits, merkle}.rs` → `ssz/tests/{bits, merkle}.rs`.
- `javm-cap/src/{cap_hash, hash, image, slot, wire}.rs` →
  `javm-cap/tests/`.

All seven only use the crate's public API. Imports rewritten from
`use super::*` / `use crate::foo::Bar` to the crate-name form
(`use ssz::*`, `use javm_cap::*`). 51 tests run cleanly across the
new integration test files.

Verified: workspace build/clippy/fmt clean; full `cargo test -p ssz`
and `cargo test -p javm-cap` green.
