+++ rust/nub-host-kvm/src/lib.rs
@@ -131,4 +131,3 @@ macro_rules! debug {
         tracing::debug!($($arg)+);
     }
 }
-
